feat: sessionlimit authenticator
Adds feature to Keycloak to restrict concurrent sessions for users.
- Based on [open upstream PR](https://github.com/keycloak/keycloak/pull/8260).
- Based on [open feature ticket](https://issues.redhat.com/browse/KEYCLOAK-849)

The reason I manually copy/pasted the changes, is, that I was unable to
build the newest version of Keycloak and unable to debug the reason for
it. Merging changes with `13.0.1` led to too many conflicts I was unable
to solve without in-depth knowledge.

These changes don't include the upstream-PR tests, as changes to the
test-suite have been made up-stream after `13.0.1`.